--- go.mod
@@ -1,26 +1,0 @@
-module github.com/BlackestDawn/nar-feed
-
-go 1.23.4
-
-require (
-	github.com/GoogleCloudPlatform/functions-framework-go v1.9.2
-	github.com/PuerkitoBio/goquery v1.10.0
-	github.com/gorilla/feeds v1.2.0
-	github.com/kardianos/service v1.2.2
-	github.com/spf13/pflag v1.0.5
-)
-
-require (
-	cloud.google.com/go/functions v1.19.3 // indirect
-	github.com/andybalholm/cascadia v1.3.2 // indirect
-	github.com/cloudevents/sdk-go/v2 v2.15.2 // indirect
-	github.com/google/uuid v1.6.0 // indirect
-	github.com/json-iterator/go v1.1.10 // indirect
-	github.com/modern-go/concurrent v0.0.0-20180228061459-e0a39a4cb421 // indirect
-	github.com/modern-go/reflect2 v0.0.0-20180701023420-4b7aa43c6742 // indirect
-	go.uber.org/atomic v1.4.0 // indirect
-	go.uber.org/multierr v1.1.0 // indirect
-	go.uber.org/zap v1.10.0 // indirect
-	golang.org/x/net v0.33.0 // indirect
-	golang.org/x/sys v0.28.0 // indirect
-)
