+++ nardata/components.go
@@ -1,4 +1,4 @@
-package data
+package nardata
 
 import (
 	"fmt"
