+++ nardata/go.mod
@@ -1,0 +1,13 @@
+module github.com/BlackestDawn/nar-feed/nardata
+
+go 1.24.2
+
+require (
+	github.com/PuerkitoBio/goquery v1.10.3
+	github.com/gorilla/feeds v1.2.0
+)
+
+require (
+	github.com/andybalholm/cascadia v1.3.3 // indirect
+	golang.org/x/net v0.39.0 // indirect
+)
